coffee-back-demo/requirements.txt
bidict==0.22.0
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.2
python-engineio==4.3.3
python-socketio==5.7.1
six==1.16.0
